+++ tools/RAiDER/models/gmao.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import h5py
 import numpy as np
-import pydap.client
+import xarray as xr
 from pyproj import CRS
 
 from RAiDER.logger import logger
@@ -80,43 +80,34 @@ class GMAO(WeatherModel):
         if corrected_DT >= T0:
             # open the dataset and pull the data
             url = 'https://opendap.nccs.nasa.gov/dods/GEOS-5/fp/0.25_deg/assim/inst3_3d_asm_Nv'
-            # For warning from pydap when using HTTPS instead of DAP2 or DAP4:
-            # pydap is incompatible with DAP data from opendap.nccs.nasa.gov. See issue #736
             with warnings.catch_warnings():
                 warnings.simplefilter("ignore")
-                ds = pydap.client.open_url(url)
+                ds = xr.open_dataset(url, decode_times=False)
 
-            q = (
-                ds['qv']
-                .array[
-                    time_ind,
-                    ml_min : (ml_max + 1),
-                    lat_min_ind : (lat_max_ind + 1),
-                    lon_min_ind : (lon_max_ind + 1),
-                ]
-                .data[0]
-            )
-            p = (
-                ds['pl']
-                .array[
-                    time_ind, ml_min : (ml_max + 1), lat_min_ind : (lat_max_ind + 1), lon_min_ind : (lon_max_ind + 1)
-                ]
-                .data[0]
-            )
-            t = (
-                ds['t']
-                .array[
-                    time_ind, ml_min : (ml_max + 1), lat_min_ind : (lat_max_ind + 1), lon_min_ind : (lon_max_ind + 1)
-                ]
-                .data[0]
-            )
-            h = (
-                ds['h']
-                .array[
-                    time_ind, ml_min : (ml_max + 1), lat_min_ind : (lat_max_ind + 1), lon_min_ind : (lon_max_ind + 1)
-                ]
-                .data[0]
-            )
+            q = ds['qv'][
+                time_ind,
+                ml_min : (ml_max + 1),
+                lat_min_ind : (lat_max_ind + 1),
+                lon_min_ind : (lon_max_ind + 1),
+            ].data
+            p = ds['pl'][
+                time_ind,
+                ml_min : (ml_max + 1),
+                lat_min_ind : (lat_max_ind + 1),
+                lon_min_ind : (lon_max_ind + 1),
+            ].data
+            t = ds['t'][
+                time_ind,
+                ml_min : (ml_max + 1),
+                lat_min_ind : (lat_max_ind + 1),
+                lon_min_ind : (lon_max_ind + 1),
+            ].data
+            h = ds['h'][
+                time_ind,
+                ml_min : (ml_max + 1),
+                lat_min_ind : (lat_max_ind + 1),
+                lon_min_ind : (lon_max_ind + 1),
+            ].data
 
         else:
             root = 'https://portal.nccs.nasa.gov/datashare/gmao/geos-fp/das/Y{}/M{:02d}/D{:02d}'
@@ -164,7 +155,7 @@ class GMAO(WeatherModel):
         self._load_model_level(f)
 
     def _load_model_level(self, filename) -> None:
-        """Get the variables from the GMAO link using OpenDAP."""
+        """Get the variables from the GMAO link using OPeNDAP."""
         # adding the import here should become absolute when transition to netcdf
         from netCDF4 import Dataset